+++ example/lib/main.dart
@@ -1,9 +1,11 @@
 import 'package:flutter/material.dart';
 import 'package:icon_shadow/icon_shadow.dart';
 
-void main() => runApp(MyApp());
+void main() => runApp(const MyApp());
 
 class MyApp extends StatelessWidget {
+  const MyApp({Key? key}) : super(key: key);
+
   // This widget is the root of your application.
   @override
   Widget build(BuildContext context) {
@@ -12,13 +14,13 @@ class MyApp extends StatelessWidget {
       theme: ThemeData(
         primarySwatch: Colors.deepPurple,
       ),
-      home: MyHomePage(),
+      home: const MyHomePage(),
     );
   }
 }
 
 class MyHomePage extends StatefulWidget {
-  MyHomePage({Key key}) : super(key: key);
+  const MyHomePage({Key? key}) : super(key: key);
 
   @override
   _MyHomePageState createState() => _MyHomePageState();
@@ -31,7 +33,7 @@ class _MyHomePageState extends State<MyHomePage> with TickerProviderStateMixin {
       backgroundColor: Colors.black54,
       appBar: AppBar(
         backgroundColor: Colors.black,
-        title: Text("Tab Bar Animation"),
+        title: const Text("Tab Bar Animation"),
       ),
       body: Center(
         child: Column(
@@ -40,7 +42,7 @@ class _MyHomePageState extends State<MyHomePage> with TickerProviderStateMixin {
             Row(
               mainAxisAlignment: MainAxisAlignment.center,
               children: <Widget>[
-                IconShadowWidget(
+                const IconShadowWidget(
                   child: Icon(
                     Icons.lightbulb_outline,
                     color: Colors.lightBlueAccent,
@@ -48,14 +50,14 @@ class _MyHomePageState extends State<MyHomePage> with TickerProviderStateMixin {
                   ),
                 ),
                 IconShadowWidget(
-                  child: Icon(
+                  child: const Icon(
                     Icons.lightbulb_outline,
                     color: Colors.lightBlueAccent,
                     size: 36,
                   ),
-                  shadowColor: Colors.lightBlueAccent.shade100,
+                  color: Colors.lightBlueAccent.shade100,
                 ),
-                IconShadowWidget(
+                const IconShadowWidget(
                   child: Icon(
                     Icons.lightbulb_outline,
                     color: Colors.lightBlueAccent,
@@ -74,7 +76,7 @@ class _MyHomePageState extends State<MyHomePage> with TickerProviderStateMixin {
                     color: Colors.green.shade900,
                     size: 36,
                   ),
-                  shadowColor: Colors.yellowAccent,
+                  color: Colors.yellowAccent,
                 ),
                 IconShadowWidget(
                   child: Icon(
@@ -97,28 +99,28 @@ class _MyHomePageState extends State<MyHomePage> with TickerProviderStateMixin {
               mainAxisAlignment: MainAxisAlignment.center,
               children: <Widget>[
                 IconShadowWidget(
-                  child: Icon(
+                  child: const Icon(
                     Icons.access_alarm,
                     color: Colors.black54,
                     size: 36,
                   ),
-                  shadowColor: Colors.yellowAccent.shade400,
+                  color: Colors.yellowAccent.shade400,
                 ),
                 IconShadowWidget(
-                  child: Icon(
+                  child: const Icon(
                     Icons.access_alarm,
                     color: Colors.black54,
                     size: 36,
                   ),
-                  shadowColor: Colors.red.shade400,
+                  color: Colors.red.shade400,
                 ),
                 IconShadowWidget(
-                  child: Icon(
+                  child: const Icon(
                     Icons.access_alarm,
                     color: Colors.black54,
                     size: 36,
                   ),
-                  shadowColor: Colors.cyanAccent.shade400,
+                  color: Colors.cyanAccent.shade400,
                 )
               ],
             ),