+++ example/lib/main.dart
@@ -1,9 +1,11 @@
 import 'package:flutter/material.dart';
 import 'package:icon_shadow/icon_shadow.dart';
 
-void main() => runApp(MyApp());
+void main() => runApp(const MyApp());
 
 class MyApp extends StatelessWidget {
+  const MyApp({Key? key}) : super(key: key);
+
   // This widget is the root of your application.
   @override
   Widget build(BuildContext context) {
@@ -12,33 +14,18 @@ class MyApp extends StatelessWidget {
       theme: ThemeData(
         primarySwatch: Colors.deepPurple,
       ),
-      home: MyHomePage(),
+      home: const MyHomePage(),
     );
   }
 }
 
-class MyHomePage extends StatefulWidget {
-  MyHomePage({Key key}) : super(key: key);
-
-  @override
-  _MyHomePageState createState() => _MyHomePageState();
-}
-
-class _MyHomePageState extends State<MyHomePage> with TickerProviderStateMixin {
-
+class MyHomePage extends StatelessWidget {
+  const MyHomePage({Key? key}) : super(key: key);
 
   @override
   Widget build(BuildContext context) {
-
-
-
-
     return Scaffold(
       backgroundColor: Colors.black54,
-      appBar: AppBar(
-        backgroundColor: Colors.black,
-        title: Text("Tab Bar Animation"),
-      ),
       body: Center(
         child: Column(
           mainAxisAlignment: MainAxisAlignment.center,
@@ -46,38 +33,92 @@ class _MyHomePageState extends State<MyHomePage> with TickerProviderStateMixin {
             Row(
               mainAxisAlignment: MainAxisAlignment.center,
               children: <Widget>[
-                IconShadowWidget(Icon(Icons.lightbulb_outline,
-                    color: Colors.lightBlueAccent, size: 36)),
-                IconShadowWidget(Icon(Icons.lightbulb_outline,
-                    color: Colors.lightBlueAccent, size: 36),shadowColor: Colors.lightBlueAccent.shade100,),
-                IconShadowWidget(Icon(Icons.lightbulb_outline,
-                    color: Colors.lightBlueAccent, size: 36),showShadow: false,)
+                const IconShadowWidget(
+                  child: Icon(
+                    Icons.lightbulb_outline,
+                    color: Colors.lightBlueAccent,
+                    size: 36,
+                  ),
+                  color: Colors.lightBlueAccent,
+                ),
+                IconShadowWidget(
+                  child: const Icon(
+                    Icons.lightbulb_outline,
+                    color: Colors.lightBlueAccent,
+                    size: 36,
+                  ),
+                  color: Colors.lightBlueAccent.shade100,
+                ),
+                const IconShadowWidget(
+                  child: Icon(
+                    Icons.lightbulb_outline,
+                    color: Colors.lightBlueAccent,
+                    size: 36,
+                  ),
+                  color: Colors.lightBlueAccent,
+                  showShadow: false,
+                )
               ],
             ),
             Row(
               mainAxisAlignment: MainAxisAlignment.center,
               children: <Widget>[
-                IconShadowWidget(Icon(Icons.home,
-                    color: Colors.green.shade900, size: 36),shadowColor: Colors.yellowAccent,),
-                IconShadowWidget(Icon(Icons.home,
-                    color: Colors.green.shade900, size: 36)),
-                IconShadowWidget(Icon(Icons.home,
-                    color: Colors.green.shade900, size: 36),showShadow: false,)
+                IconShadowWidget(
+                  child: Icon(
+                    Icons.home,
+                    color: Colors.green.shade900,
+                    size: 36,
+                  ),
+                  color: Colors.yellowAccent,
+                ),
+                IconShadowWidget(
+                  child: Icon(
+                    Icons.home,
+                    color: Colors.green.shade900,
+                    size: 36,
+                  ),
+                  color: Colors.green.shade900,
+                ),
+                IconShadowWidget(
+                  child: Icon(
+                    Icons.home,
+                    color: Colors.green.shade900,
+                    size: 36,
+                  ),
+                  showShadow: false,
+                  color: Colors.green.shade900,
+                )
               ],
             ),
-
             Row(
               mainAxisAlignment: MainAxisAlignment.center,
               children: <Widget>[
-                IconShadowWidget(Icon(Icons.access_alarm,
-                    color: Colors.black54, size: 36),shadowColor: Colors.yellowAccent.shade400,),
-                IconShadowWidget(Icon(Icons.access_alarm,
-                    color: Colors.black54, size: 36),shadowColor: Colors.red.shade400,),
-                IconShadowWidget(Icon(Icons.access_alarm,
-                    color: Colors.black54, size: 36),shadowColor: Colors.cyanAccent.shade400,)
+                IconShadowWidget(
+                  child: const Icon(
+                    Icons.access_alarm,
+                    color: Colors.black54,
+                    size: 36,
+                  ),
+                  color: Colors.yellowAccent.shade400,
+                ),
+                IconShadowWidget(
+                  child: const Icon(
+                    Icons.access_alarm,
+                    color: Colors.black54,
+                    size: 36,
+                  ),
+                  color: Colors.red.shade400,
+                ),
+                IconShadowWidget(
+                  child: const Icon(
+                    Icons.access_alarm,
+                    color: Colors.black54,
+                    size: 36,
+                  ),
+                  color: Colors.cyanAccent.shade400,
+                )
               ],
             ),
-
           ],
         ),
       ),