+++ CHANGELOG.md
@@ -1,6 +1,7 @@
  ## [2.0.0]
 
  * Add support for nullsafety
+ * Updated example
  
  ## [1.0.1]