+++ CHANGELOG.md
@@ -1,3 +1,10 @@
- ##[1.0.1] 
+ ## [2.0.0]
+
+ * Add support for nullsafety
+ * Updated example
+ * Added golden test
+ * Support parameter `offset` for the shadow
+ 
+ ## [1.0.1] 
 
 * icon shadow